+++ CHANGELOG.md
@@ -497,6 +497,7 @@ This update emphasizes improving the reliability of Predator, especially when op
 - Updated configuration back-end.
     - Increased the max-depth of the configuration validation process.
     - Predator can now automatically update the configuration file between versions when configuration values are added or removed.
+    - Improve the reliability of the configuration file location when starting Predator from an external tool, like Assassin.
 - Added an initial start-up sequence, where Predator shows some basic information before the normal start-up.
     - Predator now creates a file named `install.json` containing some basic install information on the first start-up.
 - Updated the ALPR handler.
@@ -515,3 +516,4 @@ This update emphasizes improving the reliability of Predator, especially when op
         - Previously, the time offset manager thread would keep Predator alive after using the "Quit" option.
     - All threads now watch a global variable, and will exit when it is changed.
         - This means the user can simply press Ctrl+C once, and all threads will clean up and exit.
+- Predator now saves the complete loaded alert database to the `hotlist.json` file in the interface directory.